initial structure for routes and models
I had to stub out some Koa definitions. There are some available on
flow-typed, but they are fixed to an older version of flow. I can add to
the definitions as need be.

I have a "services" directory here that contains components that require
some sort of configuration (now or in the future). This is kind of my
way of dealing with dependency injection without using some sort of
framework.

There's a lot of mixed methodologies here, specifically when it comes to
single exports versus adding onto the exports object. I usually do the
latter for babel-compiled applications because default exports can be
weird, but this is definitely node, so why not export single functions?
It doesn't matter a whole lot either way.