add web routing and scaffolding
This adds a separate nested web project and the corresponding middleware
to serve either directly from webpack (using a memory filesystem) or
from the `public` directory statically. We should be using `koa-webpack`
for the former, but there is a bug in how it handles lazy compilation
[1], so we need to write our own wrapper for now.

I think a more ideal project structure would be something like this:

    /
    + api/
    |  + app/
    |  + test/
    |  + index.js
    |  + package.json
    + web/
    |  + app/
    |  + test/
    |  + index.js
    |  + package.json
    + server/
    |  + index.js
    + index.js
    + package.json

The difference here is that we have an outer project that has a very
simple server that mounts an entire api server process on one path and
the web server process on the rest. This means that the api and web
parts can function completely independently if they needed to, and the
root project glues them together. I'm trying not to experiment with too
many new things here, so I took the half-measure of separating the web
project and using koa middleware to reach across the boundary. The
reason I still like this division is that it lets us split up the
dependencies so that, for example, we don't have `react` next to
`sequelize` in our package dependencies.

[1]: https://github.com/shellscape/koa-webpack/issues/48