add route for "authenticating"
Right now I'm doing a simple global-password auth. Authing means that if
you provide the password, you get a token that will give you access, and
that's it. I could put an expiration on the token, but it would be
arbitrary, and instead it probably makes more sense to rotate the secret
to invalidate all tokens if this ever changes.

There are some new abstractions here. `withBody` to parse bodies, since
having the middleware just add a key to the request object doesn't work
great with type-checking. Ideally there would be some way to have routes
make assertions about the request -- be they about query strings,
bodies, headers, path params, etc. -- and then those details validated
and provided explicitly. I don't know whether that should exist at the
route level (kind of like servant) or at the controller level.

I also added a `catching`, which I suppose is more of the same above
(you can kind of think about it as `withErrorHandling`). Using
`try`/`catch` everywhere is tiring, especially since `catch` cannot be
choosy about what types of errors it catches.

This also included some refactoring of the config. For one, config is
served out of `config/`, and the file used is based on your environment.
This lets us have one checked in for development purposes, and space for
a `production.env` that won't be overwritten on deploy. While working on
this, I discovered that my config library wasn't typed as well as I
thought. It turns out that flow does not understand curried arrow
functions very well, so you have to annotate the return type of each
one. I also completely mis-typed the `tree` in `fromEnv`, which started
failing once I fixed the arrow types.