fix recursion in preview generation
The poll function is recursive in that it calls the generator function
`f`, then returns a stream with 0 or more results and concatenates on a
delay plus recursive call. This works, but it produces streams like
this:

  stream [0, stream [1, stream [2, stream [3, ...

which is much like a recursive representation of a list in a functional
language. However, we need to flatten this to get out the individual
items, but that means that on item 1000, you are flattening 1000
streams, since there aren't optimizations given that you can't make
assumptions about what will be emitted on the stream.

As such, you accumulate memory and get to the point of a stack overflow
within ~1850 iterations. So this implementation switches out the
promise-based constructor with the generator constructor. We can even
still use an `async` function since generators are callback-based. One
of the benefits of the generator is that aside from pushing values, you
can use `next()` to give control to another stream, which lets us use
recursion without the flattening overhead.