fix uploads with spaces in the filename
We were passing a path to the s3 library, which meant the key would look
something like: `uploads/<hash>/file%20name.jpg', since we used the
`URL` builder to create the key. But that means the file was actually
uploaded with literal percent signs in the name. Since our s3 wrapper is
asking for a path, and since we're already modifying it to remove the
leading slash, we're going to decode the URI there.