fix back button during upload
When the upload hits 100%, that just means your request was received by
the server. If you hit back during that time, the promise will resolve
and try to `setState`, but the uploader component will have been
unmounted. There are better ways to handle this flow, but we mostly want
to have the user wait until everything hits s3.

Changing the listener from `xhr.upload` to `xhr` doesn't do much; it
just is probably the preferred option.