fix edge max-height bug
More issues with flexbox and images! It looks like the `height: 100%`
doesn't hold in Edge, since the flex items don't have heights you can
pin to. So by setting the outer container to a relative position and the
contents to an absolute position, it forces some sort of height
recalculation.