fix image resizing for all browsers
I just need to scrap `<img>`s entirely, since they are just not playing
with flexbox the way I want them to. By using a div and
`background-image`, I can let the div take up the entire width and
height and center a background image using `contain` as the sizing
property. The downside to this is that the uploader is still pushed to
the bottom and now the images will be sized *up* if possible, but these
compromises will do for now. I really want to avoid using javascript,
but it may be the only way to do it right.