do client-side mimetype checking
Instead of letting the user upload large files that aren't allowed, we
can check the mime type on the client. We're duplicating the valid mime
types since we don't have a good method of sharing static data between
the api and client.

This also brings in `babel-polyfill`, which lets us use `async/await`.