improve error reporting
We have an error list component, some better styles, and the ability to
dismiss or retry errors. There are no retryable errors in the app yet,
but we can add some to handle network blips or related timeouts.

I added a new utility to handle object keys using a `WeakMap` and
`uniqueId` behind the scene. There is no shim for `WeakMap`, so there
might be some performance consideration if you're uploading a ton on one
of those browsers. Don't use those browsers.