add uploaders model index route
This is not its own table, but is derived from the uploads table. We did
this on purpose from the beginning so that users wouldn't have to create
accounts. We are calling the "id" of the user the lowercased name and
aggregating locally on that. We could do this in the database, but right
now we have the flexibility of picking the name with the most uploads as
the display name for that identifier. Really this isn't going to matter
much and will be easy to change.

We're also changing the definition of `JSONValue` to force arrays and
objects to be immutable in that items cannot be pushed and properties
cannot be added. This makes them covariant types. This is important
because you can't pass, say, an array of strings to something that takes
a `JSONValue`, since it is possible for that function to modify the
array and add something that is otherwise a valid `JSONValue` but not a
valid `string`.