allow retrying of uploads for 5xx and network errors
I was forced to install a babel transform plugin so that I could
successfully extend `Error`. In normal ES6 classes, you can subclass
error, since you are not forced to return the result of the `super()`
call, which would override the constructor with the return value. The
babel transform, however, explicitly stores the `super()` result in
order to return it, causing the following to evaluate false:

    new NetworkError('message') instanceof NetworkError

The babel plugin writes different constructors for classes extending
specific superclasses (Error and Array by recommended configuration).

I realized that the promise from the `RetryThunk` was not very
important, since we want the retry to go through the entire enqueuing
process. Nevertheless, if an error occurs in setting up the retry, we
want to capture it and handle it.