fix scrolling on mobile
Because the stage was the entire viewport height, and because mobile
browsers have some sort of chrome that gets pushed away as you scroll,
what you see when you load the page is not the entire viewport height.
We want the gallery to fill only the visible viewport, but we also don't
want you to be able to scroll the gallery away when you push the chrome
off. So the solution is to not use `100vh`, and instead use `100%` on
the stage and all parent elements.