upgrade sequelize to 4.0
Since we don't have a huge app, there wasn't much ground to cover for
the upgrade. Ordering via raw strings is deprecated, so we switched the
one place we had used it by mistake to the array syntax. The other
change is that hooks are run by default on validations (specifically so
that `beforeValidate` and `afterValidate` hooks get run), however it
looks like this changes the return type of `validate()` unintentionally.
We're forcing hooks to false for now.