add migrations framework and initial migration
I didn't know if I was going to do this, but I don't want to regret a
botched `sync()` at some point. We're not using `sequelize-cli` since it
requires a separate configuration file, which would mean that we would
have two sets of config for connecting to postgres. No go. So we're
using umzug, the migration library that is generic but written for
Sequelize and configuring it ourselves. Turns out, it's very easy.