make effort to fix processing errors
It looks like the `gm` library has some serious troubles with streams.
Particularly, if the process errors after the initial descriptors were
acquired, or if it exits with a non-zero code, you would expect the
stream to emit an error. These cases aren't handled, and it looks like
changes have mostly stopped, so we're going to work around it a little.
Our `magickWith` helper abstracts that ugliness away. The added part
here was that if stderr is written to at all, the entire stream errors.
Though I can't say this is a beautiful solution, it might help prevent
blank images from being uploaded.

Since we have a more intermittent error case being added, this also adds
support for retrying preview generation with an exponential backoff. I
added some highland type annotations along the way.

I also took this chance to refactor some of the storage and resizer
services. Instead of taking an object describing what to do, we can just
expose two functions that use currying to get what we want.