add progress colors and fix overall total
The stats for "total files queued" was only incremented for each
`enqueueAll`, but a single enqueue could be several files, so we want to
increment by the length. There was also a minor off-by-one bug with the
upload concurrency. This adds color configuration to the progress
component as well and moves around the order of components on the
uploader. More styling would definitely help, since the horizontal space
is mostly wasted.